// src/app/features/performance-recording/growth-monitoring/models/search-response.model.ts
export interface SearchResponse {
  ownerId: number;
  animalsCount: number;
  animalsList: AnimalsList[];
  ownerUuidKey1: string;
  ownerUuidKey2: string;
  ownerName: string;
  fatherName: string;
  ownerDateOfbirth: string;
  ownerGender: string;
  ownerMobileNo: number;
  emailId: string;
  ownerAddress: string;
  ownerAddressPincode: string;
  ownerAddressStateCd: number;
  ownerStateName: string;
  ownerAddressDistrictCd: number;
  ownerDistrictName: string;
  ownerAddressCityVillageCd: number;
  ownerVillageName: string;
  ownerAddressHamletCd: string;
  ownerAddressTehsilCd: number;
  ownerTehsilName: string;
  affiliatedAgencyUnionOrPc: boolean;
  registrationStatus: string;
  villageInstitutionType: string;
  villageInstitutionTypeValue: string;
  villageInstitutionCode: string;
  villageInstitutionName: string;
  membershipNumber: string;
  modifiedDate: string;
  modifiedBy: string;
  createdBy: string;
  creationDate: string;
  locationCoordinatesLatitude: string;
  locationCoordinatesLongitude: string;
  alternateMobileNo: string;
  ownerCastCategoryCd: string;
  ownerCasteCategory: string;
  ownerLandHoldingCd: string;
  ownerLandHoldingValue: string;
  isPourerMember: boolean;
  isOwnerBelowPovertyLine: string;
  hhid: string;
  isCategoryVerified: string;
  isOwnerMobileVerified: boolean;
  orgId: string;
  orgDistrictCd: string;
  orgAddress: string;
  orgName: string;
  orgPin: string;
  orgRegistrationNo: string;
  orgType: string;
  orgStateCd: string;
  orgDistrictCdAreaOperating: string;
  orgMobileNo: string;
  districtNameAreaOperating: string;
  orgTypeDesc: string;
  panNumber: string;
  ownerTypeCd: number;
  ownerType: string;
  ownerTypeCategoryCd: number;
  ownerTypeCategory: string;
}

export interface AnimalsList {
  animalId: number;
  animalName: string;
  animalPicUrl: string;
  ageInMonths: number;
  animalStatusCd: number;
  animalStatus: string;
  breedAndExoticLevels: BreedAndExoticLevel[];
  coatColourCd: number;
  damId: string;
  damSireId: string;
  dateOfBirth: string;
  isLoanOnAnimal: boolean;
  lastDewormingDate: string;
  lastVaccinationDate: string;
  loanAmount: string;
  muzzlePicUrl: string;
  ownerId: number;
  registrationDate: string;
  registrationRemarks: string;
  registrationStatus: string;
  sex: string;
  sireId: string;
  sireSireId: string;
  speciesCd: number;
  species: string;
  tagId: number;
  taggingDate: string;
  isElite: string;
  milkingStatus: string;
  numberCalvings: string;
  pregnancyMonth: number;
  pregnancyStatus: string;
  yieldMilkRecorded: string;
  fieldSubmittedforUpdate: string;
  imagePreviewUrl: string;
  currentLactationNo: number;
  age: string;
  breed: string;
  eligibleForEt: string;
  dryOffDate: string;
  lastCalvingDate: string;
  ageInDays: string;
  animalCategory: string;
}

export interface BreedAndExoticLevel {
  breed: string;
  bloodExoticLevel: string;
}